Feature/refactor existing network layer and replace rx with combine

+++ Core/Core/Classes/Responces/CoinChartDataResponse.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 public struct CoinChartDataResponse: Decodable {
-    public let prices: [[Float]]
+    public let prices: [[Double]]
 }
